% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{bind_raw}
\alias{bind_raw}
\title{bind_raw}
\usage{
bind_raw(raw_data)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}
}
\description{
bind_raw
}
\author{
Tianjian Qin
}
