% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{bind_replication}
\alias{bind_replication}
\title{bind_replication}
\usage{
bind_replication(raw_data = NULL, nrep = NULL)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{nrep}{Number of replications}
}
\description{
bind_replication
}
\author{
Tianjian Qin
}
