% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.r
\name{bind_result}
\alias{bind_result}
\title{Binding extracted results in simulation}
\usage{
bind_result(result)
}
\arguments{
\item{result}{a list generated by edd simulation function}
}
\value{
A binded list of extracted results
}
\description{
Function to bind all results extracted from edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
