% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BuildCaretEVEstimator}
\alias{BuildCaretEVEstimator}
\title{Builds an error variance estimator using a caret model as input (caret model must have saved CV predictions)}
\usage{
BuildCaretEVEstimator(x, model, Nmax = 20, cores = 1, optFunc = CEC)
}
\arguments{
\item{x}{the training set}

\item{the}{trained caret model (must have saved CV predictions)}

\item{Nmax}{the maximum number of neighbours to consider}

\item{cores}{the number of cores to utilise during estimator building}

\item{the}{optimisation function used to score the correlation}
}
\value{
a trained error variance estimator
}
\description{
An error variance estimator is built using a caret model that was trained using cross-validation.
The \code{savePredictions} argument to the \code{trainControl} function needs to be set to TRUE
so that caret saves the cross-validation predictions during model training.
The estimator can be applied to make error estimates for
new data points using the PredictSigmas function.
}
\author{
Daniel Murrell <dsmurrell@gmail.com>
}

