% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BuildEVEstimator}
\alias{BuildEVEstimator}
\title{Builds an error variance estimator using training set, fold indexes, observed values and CV predictions}
\usage{
BuildEVEstimator(x, folds, obs, preds, Nmax = 20, cores = 1,
  optFunc = CEC)
}
\arguments{
\item{x}{the training set}

\item{folds}{the folds used in cross validation}

\item{obs}{the observed values}

\item{the}{crossvalidated predicitions}

\item{Nmax}{the maximum number of neighbours to consider}

\item{cores}{the number of cores to utilise during estimator building}

\item{the}{optimisation function used to score the correlation}
}
\value{
a trained error variance estimator
}
\description{
An error variance estimator is built using the training set, the fold indexes used during training,
the observed values of the training set and the cross-validation predictions during training.
If the cross-validation predictions are kept during training, this should allow \code{eve} to
be backwards compatible to any trained model. The estimator can be applied to make error estimates for
new data points using the PredictSigmas function.
}
\author{
Daniel Murrell <dsmurrell@gmail.com>
}

