% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_stat.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculating confidence interval for LTT plot (reversed time scale)}
\usage{
calculate_CI(brts_list, tt, alpha)
}
\description{
Calculating confidence interval for LTT plot (reversed time scale)
}
\author{
Rampal Etienne; Tianjian Qin
}
