% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_stat_utilities.R
\name{calculate_CI_alt}
\alias{calculate_CI_alt}
\title{Calculating confidence interval for LTT plot (reversed time scale)}
\usage{
calculate_CI_alt(brts_list, tt, alpha)
}
\arguments{
\item{brts_list}{A list containing branching times of a simulation}

\item{tt}{Time table}

\item{alpha}{Resolution}
}
\description{
Function to calculate confidence interval for LTT plot, its an Rcpp alternative version of the
native C++ function, used for debugging
}
\author{
Thijs Janzen; Tianjian Qin
}
