% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_result.r
\name{check_folder}
\alias{check_folder}
\title{check_folder}
\usage{
check_folder(name, verbose = TRUE)
}
\arguments{
\item{name}{A string, name of the simulation job}

\item{verbose}{Logical, decides whether to print information}
}
\description{
check_folder
}
\author{
Tianjian Qin
}
