% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_result.r
\name{check_path}
\alias{check_path}
\title{check_path}
\usage{
check_path(path, verbose = TRUE)
}
\arguments{
\item{path}{Specify the path to be checked}

\item{verbose}{Logical, decides whether to print information}
}
\description{
check_path
}
\author{
Tianjian Qin
}
