% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_utilities.R
\name{download_ftp_file}
\alias{download_ftp_file}
\title{Function to batch download files from an \code{FTP} or \code{SFTP} server.}
\usage{
download_ftp_file(file_remote, file_local, credentials = "")
}
\arguments{
\item{file_remote}{\code{URL}s of the files to be downloaded.}

\item{file_local}{File names for the local version of \code{file_remote}.
\code{download_ftp_file} will create directories if they do not exist and are
used.}

\item{credentials}{Credentials for a \code{FTP} or \code{SFTP} server. Do not
use \code{credentials} if the server does not require authentication.
\code{credentials} takes the format: \code{"username:password"}.}
}
\description{
Function to batch download files from an \code{FTP} or \code{SFTP} server.
}
\author{
Tianjian Qin
}
