% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_phylo.R
\name{draw_daughter_lineages}
\alias{draw_daughter_lineages}
\title{Draw the state history of two daughter edges connected to a node}
\usage{
draw_daughter_lineages(
  envir = parent.frame(),
  parent_node,
  phy,
  history,
  nodes,
  tolerance = 1e-09
)
}
\description{
An iterator function to look up and draw the state history of two daughter edges connected to a node
}
