% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_fetch.R
\name{edd_fetch_results}
\alias{edd_fetch_results}
\title{Function to fetch simulation results from an \code{FTP} or \code{SFTP} server.}
\usage{
edd_fetch_results(url = NULL, credentials = NULL)
}
\arguments{
\item{url}{\code{URL}s of the files to be downloaded.}

\item{credentials}{Credentials for a \code{FTP} or \code{SFTP} server. Do not
use \code{credentials} if the server does not require authentication.
\code{credentials} takes the format: \code{"username:password"}.}
}
\description{
Function to fetch simulation results from an \code{FTP} or \code{SFTP} server.
}
\author{
Tianjian Qin
}
