% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{edd_go}
\alias{edd_go}
\title{Starting a edd simulation}
\usage{
edd_go(
  nrep = 1000,
  combo = NULL,
  strategy = future::sequential,
  workers = NULL
)
}
\arguments{
\item{nrep}{number of replication}

\item{combo}{parameter space}

\item{strategy}{determine if the simulation is sequential or multi-sessioned}

\item{workers}{determine how many sessions are participated in the simulation}
}
\description{
Function to start a replicated edd simulation along with a given
parameter space
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
