% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{edd_load}
\alias{edd_load}
\title{edd_load}
\usage{
edd_load(
  path = NULL,
  strategy = "sequential",
  workers = 1,
  save_file = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Name of the simulation to be loaded}

\item{strategy}{Determine if the loading process is sequential or multi-sessioned
or multi-cored}

\item{workers}{Determine how many sessions are participated in the loading process}

\item{verbose}{Logical, decides whether to print loading details}
}
\description{
edd_load
}
\author{
Tianjian Qin
}
