% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{edd_load_split}
\alias{edd_load_split}
\title{edd_load_split}
\usage{
edd_load_split(raw_data = NULL, verbose = TRUE)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{verbose}{Logical, decides whether to print the loading details}
}
\description{
edd_load_split
}
\author{
Tianjian Qin
}
