% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{edd_merge}
\alias{edd_merge}
\title{edd_merge}
\usage{
edd_merge(path = NULL, verbose = TRUE)
}
\arguments{
\item{path}{Name of the simulation to be loaded}

\item{verbose}{Logical, decides whether to print loading details}
}
\description{
edd_merge
}
\author{
Tianjian Qin
}
