% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_branch}
\alias{edd_plot_branch}
\title{Generating boxplots of various tree branching measures}
\usage{
edd_plot_branch(
  raw_data = NULL,
  method = "treestats",
  save_plot = FALSE,
  path = NULL
)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{method}{Specify which package to be used to calculate statistics}

\item{save_plot}{Logical, whether save to file or return a ggplot object}

\item{path}{Path to save the plot}
}
\value{
an ggplot object
}
\description{
Function to generate boxplots showing the meaning branch lengths of the trees.
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
