% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_eds}
\alias{edd_plot_eds}
\title{Generating line plot of evolutionary distinctiveness of each lineage}
\usage{
edd_plot_eds(
  raw_data = NULL,
  rep_id = 1,
  save_plot = FALSE,
  path = NULL,
  deviation = FALSE,
  annotation = TRUE
)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{rep_id}{specify the id of replication to plot}

\item{save_plot}{Logical, whether save to file or return a ggplot object}

\item{path}{The path to save the plots}

\item{deviation}{Logical, whether to plot the deviation of evolutionary distinctiveness}

\item{annotation}{Logical, whether to plot the annotation of parameters}
}
\value{
an ggplot object
}
\description{
Function to generate a plot showing the transition of evolutionary
distinctiveness of each lineage
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
