% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_grouped_eds}
\alias{edd_plot_grouped_eds}
\title{Generating line plot of evolutionary distinctiveness of each lineage by group}
\usage{
edd_plot_grouped_eds(
  raw_data = NULL,
  group = "metric",
  save_plot = FALSE,
  path = NULL
)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{group}{specify the group to plot}

\item{save_plot}{Logical, whether save to file or return a ggplot object}

\item{path}{The path to save the plots}
}
\value{
an ggplot object
}
\description{
Function to generate a plot showing the transition of evolutionary distinctiveness
of each lineage by specified group
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
