% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_grouped_ltt}
\alias{edd_plot_grouped_ltt}
\title{Generating LTT plot for a replicated edd simulation}
\usage{
edd_plot_grouped_ltt(
  raw_data = NULL,
  group = "metric",
  save_plot = FALSE,
  path = NULL
)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{group}{specify the group to plot}

\item{save_plot}{Logical, whether save to file or return a ggplot object}

\item{path}{The path to save the plots}
}
\value{
an ggplot object
}
\description{
Function to generate grouped lineages through time plot from
raw data of a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
