% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_las}
\alias{edd_plot_las}
\title{Generating speciation rates plot for a replicated edd simulation}
\usage{
edd_plot_las(raw_data = NULL, rep_id = 1)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}
}
\value{
an plot containing  plot
}
\description{
Function to generate a plot showing the transition of speciation
rates from raw data of a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
