% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_ltt}
\alias{edd_plot_ltt}
\title{Generating LTT plot for a replicated edd simulation}
\usage{
edd_plot_ltt(raw_data = NULL, alpha = 0.05, save_plot = FALSE)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{alpha}{Resolution of the confidence interval}

\item{save_plot}{true or false, to decide whether to save the plots to files}
}
\value{
an plot object
}
\description{
Function to generate lineages through time plot from
raw data of a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
