% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_nltt}
\alias{edd_plot_nltt}
\title{Generating nLTT plot for a replicated edd simulation}
\usage{
edd_plot_nltt(
  raw_data = NULL,
  drop_extinct = TRUE,
  save_plot = FALSE,
  path = NULL
)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{drop_extinct}{Logical, decides whether to drop extinct lineages}

\item{save_plot}{Logical, decides whether to save the plots to files}
}
\value{
an plot object
}
\description{
Function to generate normalized lineages through time plot from
raw data of a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
