% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_temporal_dynamics}
\alias{edd_plot_temporal_dynamics}
\title{Generating plots of temporal dynamics of a single simulation}
\usage{
edd_plot_temporal_dynamics(
  raw_data = NULL,
  rep_id = 1,
  save_plot = FALSE,
  path = NULL
)
}
\arguments{
\item{raw_data}{A list of results generated by edd simulation function}

\item{rep_id}{specify the id of replication to plot}

\item{save_plot}{true or false, to decide whether to save the plots to files}

\item{path}{The path to save the plots}
}
\value{
an plot object
}
\description{
Function to generate plots of temporal dynamics of a single simulation, metrics including tree balance
indexes and tree branch related measures
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
