% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_tree}
\alias{edd_plot_tree}
\title{Plotting a tree for a selected parameter set}
\usage{
edd_plot_tree(
  raw_data = NULL,
  rep_id = stop("Please specify the id of the replicate to be plotted"),
  drop_extinct = TRUE,
  save_plot = FALSE
)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{rep_id}{the id of the replicate to be plotted}

\item{drop_extinct}{Logical, decides whether to drop extinct lineages}

\item{save_plot}{Logical, decides whether to save the plots to files}
}
\value{
an plot object
}
\description{
Function to plot a tree for a selected parameter set
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
