% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.R
\name{edd_sim_batch}
\alias{edd_sim_batch}
\title{edd_sim_batch}
\usage{
edd_sim_batch(
  combo = NULL,
  history = FALSE,
  nrep = 1000,
  strategy = "sequential",
  workers = NULL,
  verbose = FALSE
)
}
\arguments{
\item{combo}{A list containing parameter sets to be used to start a simulation}

\item{history}{Logical, decides whether to record historical states for all
of the lineages, including the transitions of the rates and the evolutionary
relationship metrics}

\item{nrep}{number of replication}

\item{strategy}{determine if the simulation is sequential or multi-sessioned
or multi-cored}

\item{workers}{determine how many sessions are participated in the simulation}

\item{verbose}{Logical, decides whether to print simulation details of each
time step during simulation, for debugging purposes}
}
\description{
edd_sim_batch
}
\author{
Tianjian Qin
}
