% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{edd_sim_rep}
\alias{edd_sim_rep}
\title{Running a replicated edd simulation}
\usage{
edd_sim_rep(nrep, pars, age, model, metric, offset)
}
\arguments{
\item{nrep}{the number of replication}

\item{pars}{parameters used in a edd simulation}

\item{age}{the total running time of a edd simulation}

\item{model}{the type of model used in a edd simulation}

\item{metric}{the metric used in a edd simulation}

\item{offset}{the offset method for Phi used in a edd simulation}
}
\value{
a list of simulation results
}
\description{
Function to automatically run a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
