% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_stat.R
\name{edd_stat}
\alias{edd_stat}
\title{Function to calculate tree statistics of simulation results}
\usage{
edd_stat(
  raw_data,
  method = "treestats",
  strategy = "sequential",
  workers = 1,
  verbose = TRUE
)
}
\arguments{
\item{raw_data}{Results of simulation}

\item{method}{Method to calculate tree statistics, default is "treestats"}

\item{strategy}{Parallel strategy, default is "sequential"}

\item{workers}{Number of workers, default is 1}

\item{verbose}{Whether to print progress, default is TRUE}
}
\value{
a data frame of tree statistics
}
\description{
Function to calculate tree statistics of simulation results
}
\author{
Tianjian Qin
}
