% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_stat.R
\name{edd_stat_rtree}
\alias{edd_stat_rtree}
\title{Function to calculate tree statistics of search_tree results}
\usage{
edd_stat_rtree(data, combo)
}
\arguments{
\item{data}{A list of trees}

\item{combo}{A list of parameter combinations}
}
\value{
a data frame of tree statistics
}
\description{
Function to calculate tree statistics of search_tree results
}
\author{
Tianjian Qin
}
