% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{edd_wrapper}
\alias{edd_wrapper}
\title{Conducting a edd simulation and plotting the results}
\usage{
edd_wrapper(input = NULL, plot_opt = NULL, stat_opt = NULL, ...)
}
\arguments{
\item{...}{other parameters passed to plot or stat functions}

\item{nrep}{the number of replication}

\item{pars}{parameters used in a edd simulation}

\item{age}{the total running time of a edd simulation}

\item{model}{the type of model used in a edd simulation}

\item{metric}{the metric used in a edd simulation}

\item{offset}{the offset method for Phi used in a edd simulation}
}
\description{
Function to automatically run a replicated edd simulation and then
produce several plots from the results
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
