% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{extract_edd_result}
\alias{extract_edd_result}
\title{Extracting a specified component of the result list generated in
simulation}
\usage{
extract_edd_result(result, nrep, which, nlist = 8)
}
\arguments{
\item{result}{A list of result generated by edd simulation function}

\item{nrep}{The number of replication used in the simulation}

\item{which}{Specifying which part of the result to be extracted}

\item{nlist}{The number of components in generated list of result}
}
\value{
A list of extracted results
}
\description{
Function to extract a specified component of the result list
generated in edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
