% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{extract_which}
\alias{extract_which}
\title{extract_which}
\usage{
extract_which(raw_data = NULL, nrep = NULL, which = NULL, nlist = 10)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{nrep}{Number of replications}

\item{which}{Which replication to extract}

\item{nlist}{Number of lists contained in a list of results}
}
\description{
extract_which
}
\author{
Tianjian Qin
}
