% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logs.r
\name{job_detail}
\alias{job_detail}
\title{Parsing log files in a folder into a list of job details}
\usage{
job_detail(path = NULL)
}
\arguments{
\item{path}{path to the folder containing the log files}
}
\description{
Function to parse job details in all log files of a folder, then
return a list storing those information
}
\author{
Tianjian Qin
}
