% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logs.r
\name{job_parameter}
\alias{job_parameter}
\title{Find out which parameter set is used for a job}
\usage{
job_parameter(path = NULL, which = NULL)
}
\arguments{
\item{path}{path to the result folder}

\item{which}{specify which job to be parsed}
}
\description{
Function to find out which parameter set is used for a job
}
\author{
Tianjian Qin
}
