% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logs.R
\name{job_status}
\alias{job_status}
\title{Parsing log files in a folder into a matrix of job status}
\usage{
job_status(path = NULL)
}
\arguments{
\item{path}{path to the folder containing the log files}
}
\description{
Function to parse outputs in all log files of a folder, then
return a matrix to show the status of the jobs
}
\author{
Tianjian Qin
}
