% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{match_raw}
\alias{match_raw}
\title{Matching lineages and values in raw data}
\usage{
match_raw(raw_data = NULL, id)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{nrep}{the number of replications}
}
\value{
A matched data frame of the results
}
\description{
Function to matach lineages names and corresponding values from
replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
