% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{match_time_step}
\alias{match_time_step}
\title{match_time_step}
\usage{
match_time_step(raw_data = NULL, hist_state = NULL)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{hist_state}{Match time step to historical state matrices}
}
\description{
match_time_step
}
\author{
Tianjian Qin
}
