% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{match_which}
\alias{match_which}
\title{match_which}
\usage{
match_which(raw_data = NULL, which = NULL)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{which}{Which part of data to be matched}
}
\description{
match_which
}
\author{
Tianjian Qin
}
