% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{plot_eds}
\alias{plot_eds}
\title{Generating plot files of evolutionary distinctiveness-es}
\usage{
plot_eds(raw_data)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}
}
\description{
Function to generate plot files showing the transition of
evolutionary distinctiveness of each lineage
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
