% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{plot_las}
\alias{plot_las}
\title{Generating plot files of speciation rates}
\usage{
plot_las(raw_data)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}
}
\description{
Function to generate plot files showing the transition of
speciation rate of each lineage
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
