% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{plot_ltt}
\alias{plot_ltt}
\title{Generating plot files of lineages-through-time with confidence interval}
\usage{
plot_ltt(raw_data)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}
}
\description{
Function to generate plot files showing the lineages-through-time
curves with confidence intervals
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
