% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{plot_mus}
\alias{plot_mus}
\title{Generating plot files of extinction rates}
\usage{
plot_mus(raw_data)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}
}
\description{
Function to generate plot files showing the transition of
extinction rate of each lineage
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
