% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PredictSigmas}
\alias{PredictSigmas}
\title{Predict the error variance of new data points given the dataset and a trained estimator}
\usage{
PredictSigmas(x, estimator)
}
\arguments{
\item{x}{the descriptors of the new datapoints transformed in exactly the same way as they were during training}

\item{the}{trained caret model (must have saved CV predictions)}

\item{Nmax}{the maximum number of neighbours to consider}

\item{cores}{the number of cores to utilise during estimator building}

\item{the}{optimisation function used to score the correlation}
}
\value{
the predicted error variances for the new data points
}
\description{
Given new datapoints, \code{PredictSigmas} used a trained error variance estimator to predict the
sigmas of the error distribution for those datapoints. It is important that the descriptors are
transformed in exactly the same way as the descriptors were transformed before model training.
}
\author{
Daniel Murrell <dsmurrell@gmail.com>
}

