% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{reduce_element}
\alias{reduce_element}
\title{reduce_element}
\usage{
reduce_element(ls, ...)
}
\arguments{
\item{ls}{The list to drop some elements}

\item{...}{Names of the elements to be dropped from the list}
}
\description{
Reduce specified element(s) from a list
}
\author{
Tianjian Qin
}
