% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_rtree.R
\name{search_tree}
\alias{search_tree}
\title{Function to search for trees with a given number of tips and a given parameter set}
\usage{
search_tree(
  pars,
  age,
  model,
  metric,
  offset,
  size = NULL,
  n = 1,
  history = FALSE,
  verbose = FALSE,
  size_limit = 1e+06
)
}
\arguments{
\item{pars}{Vector of parameters: \cr \cr \code{pars[1]} corresponds to
lambda (speciation rate) \cr \code{pars[2]} corresponds to mu (extinction
rate) \cr \code{pars[3]} corresponds to beta_num (coefficient for species
number effect on speciation) \cr \code{pars[4]} corresponds to beta_phi
(coefficient for evolutionary distinctiveness effect on speciation)
\cr \code{pars[5]} corresponds to gamma_num (coefficient for species number
effect on speciation) \cr \code{pars[6]} corresponds to gamma_phi
(coefficient for evolutionary distinctiveness effect on extinction)}

\item{age}{Sets the crown age for the simulation}

\item{model}{Sets the model of diversity-dependence: \cr \code{model ==
dsce2} : linear dependence in speciation rate with parameters beta_num and
beta-phi\cr \code{model == dsde2} : linear dependence
in both speciation rate and extinction rate with parameters beta_num,
beta_phi, gamma_num and gamma_phi}

\item{metric}{"pd" , "ed" or "nnd", Specifies which evolutionary distinctiveness
metric should be used.}

\item{offset}{Specifies which method to use to offset the impact of tree age
and the collinearity between pd and species richness. "none" for no offset
method; "simtime" for deducting tree age from pd value; "spcount" for
dividing pd value by species richness; "both" for applying both "simtime" and
"spcount", by firstly deducting tree age and then dividing by species richness}

\item{size}{Specifies the number of tips in the tree to be searched for}

\item{n}{Specifies the number of trees to be searched for}

\item{history}{Logical, indicating whether to record the historical states
(of the rates and ED/PD values)}

\item{verbose}{Logical, for debugging purpose, indicating whether to print
simulation info at each step in the console, and save running time to a file}

\item{size_limit}{Specifies the maximum number of tips in the tree to be}

\item{converter}{Either "cpp" or "r", choose which version of L2phylo to use.}
}
\value{
a list of trees
}
\description{
Function to search for trees with a given number of tips and a given parameter set
}
\author{
Tianjian Qin
}
