% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_phylo.R
\name{stat_histree}
\alias{stat_histree}
\title{Mapping historical states to a phylo tree}
\usage{
stat_histree(phy, history)
}
\arguments{
\item{phy}{A phylo object, the tree to map historical states to}

\item{history}{A data frame containing time steps and the historical states of each lineage. The first row of the data
frame should be time, which has each time step as a row. The other rows should be named t1, t2, t3, etc, and the names
should be in the same order as they speciated. The rows should contain the values of its historical states at each time
step. At each time step where a lineage is not present, the value should be NA.}
}
\value{
A data frame of the segments of the tree
}
\description{
Function to map historical states to a phylo tree and output a data frame of segments of the tree
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
