% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{swap_layers}
\alias{swap_layers}
\title{swap_layers}
\usage{
swap_layers(ls)
}
\description{
Swap the first and the second layers of a list of lists
This function is superseded by purrr::transpose() which is way faster
}
\author{
Tianjian Qin
}
