% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{bind_raw}
\alias{bind_raw}
\title{Binding extracted results in simulation}
\usage{
bind_raw(raw_data = NULL, nrep = NULL)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{nrep}{the number of replications}
}
\value{
A binded list of the results
}
\description{
Function to bind all results extracted from replicated edd
simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
