% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_stat_utilities.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculating confidence interval for LTT plot (reversed time scale)}
\usage{
calculate_CI(brts_list, tt, alpha)
}
\arguments{
\item{brts_list}{A list containing branching times of a simulation}

\item{tt}{Time table}

\item{alpha}{Resolution}
}
\description{
Calculating confidence interval for LTT plot (reversed time scale)
}
\author{
Thijs Janzen; Tianjian Qin
}
