% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{edd_combo_maker}
\alias{edd_combo_maker}
\title{Making a parameter space for edd simulation}
\usage{
edd_combo_maker(save_file = FALSE, ...)
}
\arguments{
\item{save_file}{Logical, decides whether to save parameter sets to file}

\item{...}{parameters that form a parameter space}
}
\description{
Function to conveniently make a parameter space for edd
simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
