% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{edd_go}
\alias{edd_go}
\title{Starting a edd simulation}
\usage{
edd_go(
  combo = NULL,
  history = FALSE,
  nrep = 1000,
  name = NULL,
  seed = NULL,
  strategy = "sequential",
  workers = NULL,
  verbose = FALSE
)
}
\arguments{
\item{combo}{A list containing parameter sets to be used to start a simulation}

\item{history}{Logical, decides whether to record historical states for all
of the lineages, including the transitions of the rates and the evolutionary
relationship metrics}

\item{nrep}{number of replication}

\item{name}{name of the folder to save results, if left blank, a folder will
be created according to current time and date. Set it to "no_save" to disable
result saving function, an object containing all the results will be returned
instead.}

\item{seed}{set random seed}

\item{strategy}{determine if the simulation is sequential or multi-sessioned
or multi-cored}

\item{workers}{determine how many sessions are participated in the simulation}

\item{verbose}{Logical, decides whether to print simulation details of each
time step during simulation, for debugging purposes}
}
\description{
Function to start a replicated edd simulation along with a given
parameter space
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
