% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.R
\name{edd_plot}
\alias{edd_plot}
\title{Generating plots for a replicated edd simulation}
\usage{
edd_plot(
  raw_data = NULL,
  which = "all",
  save_plot = FALSE,
  strategy = "sequential",
  workers = 1,
  verbose = TRUE
)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{which}{Which part of data to be plotted}

\item{save_plot}{Logical, decides whether to save the plots to files}

\item{strategy}{Determine if the simulation is sequential or multi-sessioned
or multi-cored}

\item{workers}{Determine how many sessions are participated in the simulation}

\item{verbose}{Logical, decides whether to print loading details}
}
\value{
a plot pack containing several plots
}
\description{
Function to automatically generate several plots from raw data
of a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
