% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot}
\alias{edd_plot}
\title{Generating plots for a replicated edd simulation}
\usage{
edd_plot(raw_data = NULL, save_plot = FALSE)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{save_plot}{Logical, decides whether to save the plots to files}
}
\value{
a plot pack containing several plots
}
\description{
Function to automatically generate several plots from raw data
of a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
