% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_las}
\alias{edd_plot_las}
\title{Generating line plot of speciation rates of each lineage}
\usage{
edd_plot_las(raw_data = NULL, rep_id = 1, save_plot = FALSE, path = NULL)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{rep_id}{specify the id of replication to plot}

\item{save_plot}{Logical, whether save to file or return a ggplot object}
}
\value{
an ggplot object
}
\description{
Function to generate a plot showing the transition of speciation
rates of each lineage
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
