% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_plot.r
\name{edd_plot_nltt}
\alias{edd_plot_nltt}
\title{Generating nltt plot for a replicated edd simulation}
\usage{
edd_plot_nltt(raw_data = NULL, plot_options = NA, ...)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}

\item{tr}{true of false, to decide whether to truncate the plot to the range
between 0 and simulation age}
}
\value{
an nltt plot
}
\description{
Function to generate lineages through time plot from raw data
of a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
