% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{edd_sim_batch}
\alias{edd_sim_batch}
\title{Conducting a edd simulation and plotting the results}
\usage{
edd_sim_batch(
  nrep = 1000,
  combo = NULL,
  make_plot = FALSE,
  make_stat = FALSE,
  plot_opt = NULL,
  stat_opt = NULL
)
}
\arguments{
\item{nrep}{the number of replication}

\item{pars}{parameters used in a edd simulation}

\item{age}{the total running time of a edd simulation}

\item{model}{the type of model used in a edd simulation}

\item{metric}{the metric used in a edd simulation}

\item{offset}{the offset method for Phi used in a edd simulation}

\item{...}{other parameters passed to plot or stat functions}
}
\description{
Function to automatically run a replicated edd simulation and then
produce several plots from the results
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
