% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{edd_sim_batch}
\alias{edd_sim_batch}
\title{edd_sim_batch}
\usage{
edd_sim_batch(
  nrep = 1000,
  combo = NULL,
  strategy = future::sequential,
  workers = NULL
)
}
\value{

}
\description{
edd_sim_batch
}
\author{
Tianjian Qin
}
