% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{edd_sim_batch}
\alias{edd_sim_batch}
\title{edd_sim_batch}
\usage{
edd_sim_batch(
  combo = NULL,
  history = FALSE,
  nrep = 1000,
  strategy = "sequential",
  workers = NULL,
  verbose = FALSE
)
}
\description{
edd_sim_batch
}
\author{
Tianjian Qin
}
