% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_stat.R
\name{edd_stat}
\alias{edd_stat}
\title{Generating plots for a replicated edd simulation}
\usage{
edd_stat(raw_data = NULL)
}
\arguments{
\item{raw_data}{a list of results generated by edd simulation function}
}
\value{
a plot pack containing several plots
}
\description{
Function to automatically generate several plots from raw data
of a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
