% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{eve}
\alias{eve}
\title{Equivariant Variance Estimator (EVE)}
\usage{
eve(x, K = NULL, Kmax = 20)
}
\arguments{
\item{x}{Numeric vector. Input time vector.}

\item{K}{Integer (or vector of integers). The length of the \eqn{Y} vector
to be used in estimating the variance. If \code{NULL}, \eqn{K} is selected
automatically. Default: \code{NULL}.}

\item{Kmax}{Integer. Upper bound for the search when \code{K = NULL}. Default: 20.}
}
\value{
A list with:
\itemize{
\item \code{sigmahat} — Named vector of estimated standard deviation(s) using EVE wits selected \eqn{K}s.
\item \code{K} — The specified or automatically selected value(s) of \eqn{K}.
}
}
\description{
\code{eve} computes the Equivariant Variance Estimator (EVE) for a time
series under frequent mean changes. The method is based on fitting a linear
regression to the sequence \eqn{Y_k = T_k/2}, where \eqn{T_k} is constructed
from shifted squared differences of the input series. The variance estimate
is given by the intercept of the fitted regression.
}
\details{
Users may specify a fixed parameter \eqn{K}; alternatively, if
\code{K = NULL}, the optimal \eqn{K} is selected automatically using a
data-driven procedure with maximum search bound \code{Kmax}.

When \code{K = NULL}, \code{eve} internally calls \code{eve_sel_K()} to select
\eqn{K} using the standardized prediction error criterion. When \code{K} is
provided, regression is performed separately for each value of \eqn{K}.
A warning is issued if the estimated variance is negative.
}
\examples{
set.seed(2023)
n <- 1000
mu <- rep(c(rep(5, 50), rep(0, 50)), length.out = n)
eps <- rnorm(n, sd=2)
x <- eps + mu # true standard deviation is 1

sd(eps)

eve(x)              # EVE with automatic selection of K
eve(x, K = 5)       # EVE with user–specified K
eve(x, K = c(5,10,15,20)) # Compare several choices of K


}
\references{
Hao, N., Niu, Y. S., & Xiao, H. (2023).
\emph{Equivariant variance estimation for multiple change-point model}.
Electronic Journal of Statistics, 17(2), 3811–3853. arXiv:2108.09431.
}
