% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{extract_edd_result}
\alias{extract_edd_result}
\title{extract_edd_result}
\usage{
extract_edd_result(result, nrep, which, nlist = 8)
}
\value{

}
\description{
extract_edd_result
}
\author{
Tianjian Qin
}
