% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_load.R
\name{match_raw}
\alias{match_raw}
\title{match_raw}
\usage{
match_raw(x = NULL, y = NULL)
}
\arguments{
\item{x}{Data to be matched}

\item{y}{List of lineages at all the time steps}
}
\description{
match_raw
}
\author{
Tianjian Qin
}
