% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_wrapper.r
\name{match_raw}
\alias{match_raw}
\title{Matching lineages and values in raw data}
\usage{
match_raw(x = NULL, y = NULL)
}
\arguments{
\item{x}{a list of values to be matched with lineages}

\item{y}{the lineage lists of the replicated edd simulation}
}
\value{
A matched list with values and corresponding lineages
}
\description{
Function to match lineages names and corresponding values from
replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
